Detect native stream keypress events and bail.
This is because blessed is running its own version of
readline's `emitKeypressEvents()` to handle mouse escape
codes. Both implementations can be added to the same stream,
resulting in double keypress events.

Refusing to add the handler if another is present does not
fix the mouse escape codes, but it will warn and fix the
double input.